SkyreaderGuildServer/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
python-jose[cryptography]>=3.3.0
pytest>=8.0.0
httpx>=0.27.0
